+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Added
 
+- `Select`, `SelectMulti` and `InputSearch` now support `isLoading` and `detail` prop on options
+- `Combobox` and `ComboboxMulti` `openOnClick` prop
+- `ComboboxInput` now supports `freeInput` prop
 - `Span` - same as `Text` without the annoying `fontSize="medium"` (inherits by default instead)
 - `Truncate` component
 - Visual Snapshot test for `MenuItem` and `MenuGroup`
@@ -23,6 +26,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Tooltip` now has a default `maxWidth` of `30rem` (this can be overridden)
 - `Span` replaced all library-internal usage of `Text` with `Span`
 - `Code`, `CodeBlock` & `Paragraph` now explicitly use `theme.colors.text` as default color
+
+### Changed
+
+- `InputSearch` `onChange` callback argument is now a string rather than an event
+
+### Fixed
+
+- `ComboboxMulti` issue with chips not updating reflecting updated option labels
+
+### Removed
+
+- `InputSearch` `onClear` – use `onChange` with a check for empty value instead
+- `InputSearch` and `InputChips` `hideControls` – use `isClearable={false}` instead
+
+### Preview / Experimental
+
+- Experimental: `InputSearch` supports `onSelectOption`, `changeOnSelect`, `clearOnClose` and all `Select` props except `isFilterable`, `onFilter`, `showCreate` and `formatCreateLabel`
 
 ## [0.9.16] - 2020-10-02
 
@@ -98,12 +118,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Select` overwriting search value with the current option value if the option's value and label are different
 - Storybook fixes to silence console warnings
 
-### Remove
-
-- `Tooltip` & `Popover` no longer support (optional) arrow indicator
-
 ### Removed
 
+- `Tooltip` & `Popover` no longer support (optional) arrow indicator
 - `prismTheme` is no longer published
 - `theme.colors.palette` is no longer available (`palette` _can_ be import from `@looker/design-tokens` but this is a legacy behavior and is not encouraged)